feat: localize credentials struct
