+++ crates/context/config/src/client/near.rs
@@ -5,7 +5,7 @@ use std::collections::BTreeMap;
 use std::time;
 
 pub use near_crypto::SecretKey;
-use near_crypto::{InMemorySigner, Signer};
+use near_crypto::{InMemorySigner, PublicKey, Signer};
 use near_jsonrpc_client::methods::query::{RpcQueryRequest, RpcQueryResponse};
 use near_jsonrpc_client::methods::send_tx::RpcSendTransactionRequest;
 use near_jsonrpc_client::methods::tx::RpcTransactionStatusRequest;
@@ -22,10 +22,75 @@ use near_primitives::views::{
     AccessKeyPermissionView, AccessKeyView, CallResult, FinalExecutionStatus, QueryRequest,
     TxExecutionStatus,
 };
+use serde::{Deserialize, Serialize};
 use thiserror::Error;
 use url::Url;
 
 use super::{Operation, Transport, TransportRequest};
+
+#[derive(Clone, Debug, Deserialize, Serialize)]
+#[serde(try_from = "serde_creds::Credentials")]
+pub struct Credentials {
+    pub account_id: AccountId,
+    pub public_key: PublicKey,
+    pub secret_key: SecretKey,
+}
+
+mod serde_creds {
+    use near_crypto::{PublicKey, SecretKey};
+    use near_primitives::types::AccountId;
+    use serde::{Deserialize, Serialize};
+
+    #[derive(Debug, Deserialize, Serialize)]
+    pub struct Credentials {
+        account_id: AccountId,
+        public_key: PublicKey,
+        secret_key: SecretKey,
+    }
+
+    impl TryFrom<Credentials> for super::Credentials {
+        type Error = &'static str;
+
+        fn try_from(creds: Credentials) -> Result<Self, Self::Error> {
+            'pass: {
+                if let SecretKey::ED25519(key) = &creds.secret_key {
+                    let mut buf = [0; 32];
+
+                    buf.copy_from_slice(&key.0[..32]);
+
+                    if ed25519_dalek::SigningKey::from_bytes(&buf)
+                        .verifying_key()
+                        .as_bytes()
+                        == &key.0[32..]
+                    {
+                        break 'pass;
+                    }
+                } else if creds.public_key == creds.secret_key.public_key() {
+                    break 'pass;
+                }
+
+                return Err("public key and secret key do not match");
+            };
+
+            if creds.account_id.get_account_type().is_implicit() {
+                let Ok(public_key) = PublicKey::from_near_implicit_account(&creds.account_id)
+                else {
+                    return Err("fatal: failed to derive public key from implicit account ID");
+                };
+
+                if creds.public_key != public_key {
+                    return Err("implicit account ID and public key do not match");
+                }
+            }
+
+            Ok(Self {
+                account_id: creds.account_id,
+                public_key: creds.public_key,
+                secret_key: creds.secret_key,
+            })
+        }
+    }
+}
 
 #[derive(Debug)]
 pub struct NetworkConfig {
